\name{ClinData}
\docType{data}
\alias{ClinData}
\title{Example of clinical data}
\description{
An example of clinical data.
}
\usage{
data(ClinData)
}
\details{
The example data was derived from IMvigor210CoreBiologies, which includes RNA-seq data of 348 samples and corresponding drug response of atezolizumab.
}
\references{
Mariathasan S, Turley S J, Nickles D, et al. TGFbeta attenuates tumour response to PD-L1 blockade by contributing to exclusion of T cells. Nature. 2018, 554(7693): 544-8.
}
\examples{
data(ClinData)
head(ClinData)
}
\keyword{datasets}

