% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{condiGRN}
\alias{condiGRN}
\title{Build conditional GRN}
\usage{
condiGRN(exp, tf2tar, method , pValue, threshold)
}
\arguments{
\item{exp}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The TF-target interaction pairs within a prior GRN.}

\item{method}{The method used,such as 'Boruta', 'RGBM'.}

\item{pValue}{Confidence level used in Boruta.}

\item{threshold}{The threshould for weight in RGBM.}
}
\value{
A conditional GRN.
}
\description{
Build conditional gene regulatory network (GRN) with expression data and prior network by using feature selection algorithm.
}
\details{
While using Boruta, the threshold is the predifined pValue. While using RGBM or GENIE, users could set threshould based on output of weight.
}
\examples{
\donttest{
# Build a conditional GRN based on a prior GRN.
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

data(clin_data)
head(clin_data)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]
temp <- exp.1[sample(1:nrow(exp.1),1000),]

## using method Boruta
net.1 <- condiGRN(exp = temp, tf2tar = tf2tar, method = 'Boruta', pValue = 0.01)

}

}
