\name{DE}
\docType{data}
\alias{DE}
\title{ Example of DEGs input for function DysReg and DiffRegPlus}
\description{
An example of DEGs input for function DysReg and DiffRegPlus
}
\usage{
data(DE)
}
\details{
de.genes contains three columns, GeneSymbol is the gene symbol of DEGs, high.condition indicates which comdition expresses high level, de.logFC is logFC value.
}
\examples{
data(DE)
head(DE)
}
\keyword{datasets}

