% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{DiffCor}
\alias{DiffCor}
\title{Differential correlation analysis}
\usage{
DiffCor(exp.1, exp.2, tf2tar, cor.method = 'pearson', p.adj)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The TF-target interaction pairs within a prior GRN.}

\item{cor.method}{Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default) or "spearman", can be abbreviated.}

\item{p.adj}{Correction method for p value adjust.}
}
\value{
The identified differential correlation regulations.
}
\description{
Differential correlation analysis with Fishers' Z test.
}
\examples{
\donttest{

data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]

group.2 <- clin_data$sample[which(clin_data$ResponseLevel \%in\% c('CR','PR','SD'))]
exp.2 <- exp[test.genes,colnames(exp) \%in\% group.2]

dc.res <- DiffCor(exp.1, exp.2, tf2tar, 
                  cor.method = 'pearson', p.adj = 'BH')
                   
dc.res <- dc.res[dc.res$p.adj < 0.05,]
head(diff.res)
}

}
