% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{DiffCorPlus}
\alias{DiffCorPlus}
\title{Differential correlation analysis plus differential expression of target, and the consistency between differential correlation and differential expression.}
\usage{
DiffCorPlus(exp.1, exp.2, tf2tar, de.genes, cor.method = 'pearson', p.adj,
                   de.pval=NULL, de.qval = NULL, de.logFC = NULL)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The TF-target interaction pairs within a prior GRN.}

\item{de.genes}{A dataframe for differential expression genes. The dataframe must include three colomns, "GeneSymbol", "high.condition", "logFC". "high.condition" means which condition represents high expression. "logFC" is the output logFC from differential expression analysis.}

\item{cor.method}{Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default) or "spearman", can be abbreviated.}

\item{p.adj}{Correction method for p value adjust.}

\item{de.pval}{The cutoff of pval used for filtering differential expression genes.}

\item{de.qval}{The cutoff of qval used for filtering differential expression genes.}

\item{de.logFC}{The cutoff of absolute logFC used for filtering differential expression genes.}
}
\value{
The identified regulations.
}
\description{
Identify regulations with differential correlation analysis, differential expression of target, and the consistency between differential correlation and differential expression.
}
\examples{
\donttest{
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]

group.2 <- clin_data$sample[which(clin_data$ResponseLevel \%in\% c('CR','PR','SD'))]
exp.2 <- exp[,colnames(exp) \%in\% group.2]

diffcorpp.res <- DiffCorPlus(exp.1,exp.2, tf2tar, de.genes = NULL, 
                             cor.method = 'pearson', p.adj = 'BH',
                             de.pval=0.05)

diffcorpp.res <- diffcorpp.res[diffcorpp.res$p.val < 0.05,]
head(diffcorpp.res)
}

}
