% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{DiffReg}
\alias{DiffReg}
\title{Differential regulation analysis}
\usage{
DiffReg(exp.1, exp.2, tf2tar, cor.method = 'pearson', p.adj)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The  prior reference GRN containing TF-target interaction pairs.}

\item{cor.method}{The method used to calculate correlation coefficient.}

\item{p.adj}{Correction method for p value adjust.}
}
\value{
The identified differential regulations.
}
\description{
Differential regulation analysis with Fishers' Z test.
}
\examples{
\donttest{

data(ExpData) 
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]

set.seed(1234)
test.genes <- sample(1:nrow(ExpData),1000)

tmp.1 <- exp.1[test.genes,]
tmp.2 <- exp.2[test.genes,]

## implement differential regulation analysis
diffreg.res <- DiffReg(tmp.1, tmp.2, tf2tar, 
                       cor.method = 'pearson', p.adj = 'BH')
                       
## set cutoff
diffreg.res <- subset(diffreg.res,p.val < 0.01)
head(diffreg.res)

}

}
