% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{DysReg}
\alias{DysReg}
\title{Identify gene dysregulations}
\usage{
DysReg(exp.1, exp.2, tf2tar, 
       de.genes = NULL, de.pval = NULL, de.qval = NULL, de.logFC = NULL, 
       grn.method = 'Boruta', pValue = 0.01, threshold = NULL, 
       ci = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The  prior reference GRN containing TF-target relationships.}

\item{de.genes}{A dataframe for differential expression genes. If no de.genes offered, DysReg uses the default method limma to implement differential expression analysis. If de.genes offered, The dataframe must include three columns, "GeneSymbol", "high.condition", "de.logFC". "high.condition" means which condition represents high expression level. "de.logFC" is the output logFC from differential expression analysis.}

\item{de.pval}{The cutoff of pval for filtering differential expression genes. If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.05.}

\item{de.qval}{The cutoff of qval used for filtering differential expression genes.  If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.05.}

\item{de.logFC}{The cutoff of absolute logFC used for filtering differential expression genes. If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.5. This parameter could be used by combining with de.pval or de.qval.}

\item{grn.method}{The method used to build conditional GRN, such as 'Boruta', or 'RGBM'.}

\item{pValue}{Confidence level used in Boruta. Default value should be used.}

\item{threshold}{The threshould for weight in RGBM.}

\item{ci}{The confidence invetal of coefficient.}

\item{verbose}{A logical value indicating whether display the computating progress.}

\item{...}{Other parameters passed to \code{\link{condiGRN}}.}
}
\value{
The results of gene dysregulation analysis:
 \item{de.genes}{The differential expression genes between conditions.}
 \item{dysreg}{The identified gene dysregulations}
}
\description{
Identify gene dysregulations by integrating three properties including differential regulation, differential expression of target, and the consistency between differential regulation and differential expression. DysReg could consider the combinational effect of multiple regulators to target expresion in gene dysregulation analysis.
}
\details{
DysReg first builds conditional GRNs with feature selection algorithm, where regulatory intensity and its confidence interval of each link is estimated by a de-biased LASSO method. Gene dysregulations were then identified by integrating three properties including differential regulation, differential expression of target, and the consistency between differential regulation and differential expression.
}
\examples{
\donttest{
data(ExpData)
ExpData[1:5,1:5]

data(tf2tar)
head(tf2tar)

data(ClinData)
head(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]

dysreg.out <- DysReg(exp.1 = exp.1, exp.2 = exp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', 
                     pValue = 0.01, ci = 0.90, verbose = T)

dysreg <- dysreg.out$dysreg
head(dysreg)

}

}
\references{
Li Q, Li J, Dai W, et al. Differential regulation analysis reveals dysfunctional regulatory mechanism involving transcription factors and microRNAs in gastric carcinogenesis. Artif Intell Med. 2017, 77, 12-22.
}
\seealso{
\code{\link{condiGRN}}; \code{\link{quantiReg}}
}
