\name{ExpData}
\docType{data}
\alias{ExpData}
\title{Example of expression data}
\description{
An example of expression data.
}
\usage{
data(ExpData)
}
\details{
The example data was derived from IMvigor210CoreBiologies, which includes RNA-seq data of 348 samples and corresponding drug response of atezolizumab.
}
\examples{
data(ExpData)
dim(ExpData)
ExpData[1:5,1:5]
}
\keyword{datasets}
