% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{fitness.AUC}
\alias{fitness.AUC}
\title{Calculate AUC in classifying data for using as fitness function}
\usage{
fitness.Cindex(combine.dysreg, dysreg, exp.data, pheno.data, train.rate)
}
\arguments{
\item{combine.dysreg}{The combination of genes.}

\item{dysreg}{The dysregulations output from \code{\link{DysReg}}.}

\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{pheno.data}{The phenotype data corresponding to objective function. At here, it is classifying data.}

\item{train.rate}{The rate of sample for training model in cross-validation.}
}
\value{
The fitness, AUC, for each individual.
}
\description{
Calculate AUC with cross-validation in classifying data for using as fitness function.
}
\seealso{
\code{\link{combineDysreg}}
}
