% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{fitness.Cindex}
\alias{fitness.Cindex}
\title{Calculate C-Index within survival data for using as fitness function}
\usage{
fitness.Cindex(combine.dysreg, dysreg, exp.data, pheno.data, train.rate)
}
\arguments{
\item{combine.dysreg}{The combination of dysregulations.}

\item{dysreg}{The dysregulations output from \code{\link{DysReg}}.}

\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{pheno.data}{The phenotype data corresponding to fitness function. At here, it is survival data.}

\item{train.rate}{The rate of sample for training model in cross-validation.}
}
\value{
The fitness, C-Index, for each individual.
}
\description{
Calculate C-Index with cross-validation in survival data for using as fitness function.
}
\seealso{
\code{\link{combineDysreg}}
}
