% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{Obj.Cindex}
\alias{Obj.Cindex}
\title{Calculate C-Index in survival data for using as Objective function.}
\usage{
Obj.Cindex(combine.genes, dysreg, exp.data, 
                  pheno.data, train.rate)
}
\arguments{
\item{combine.genes}{The combination of genes.}

\item{dysreg}{The dysregulations output from \code{DysReg}.}

\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{pheno.data}{The phenotype data corresponding to objective function. At here, it is survival data.}

\item{train.rate}{The rate of sample for training model in cross-validation.}
}
\value{
The fitness, C-Index, for each individual.
}
\description{
Calculate C-Index with cross-validation in survival data for using as Objective function.
}
