% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{plotDysregExp}
\alias{plotDysregExp}
\title{Visualize expression pattern of a dysreglaiton between conditions}
\usage{
plotDysregExp(tf, tar, exp.1, exp.2, exp1.lab, exp2.lab,
                      method, dysreg, conf.int.level, ...)
}
\arguments{
\item{tf}{The TF of a gene dysregulation.}

\item{tar}{The target of a dysregulation.}

\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp1.lab}{The label condition of exp.1.}

\item{exp2.lab}{The label condition of exp.2.}

\item{method}{The mehod of dysregulation analysis, "dysreg", "diffcorplus", or "diffcor".}

\item{dysreg}{The results of dysreg.}

\item{conf.int.level}{Level controlling confidence region.}

\item{...}{further arguments to \code{ggscatter}.}
}
\description{
Visualize expression pattern of a dysreglaiton between conditions.
}
\examples{
\donttest{
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]

group.2 <- clin_data$sample[which(clin_data$ResponseLevel \%in\% c('CR','PR','SD'))]
exp.2 <- exp[test.genes,colnames(exp) \%in\% group.2]

set.seed(1234)
test.genes <- sample(1:nrow(exp),1000)

temp.1 <- exp.1[test.genes,]
temp.2 <- exp.2[test.genes,]

dysreg.out <- DysReg(exp.1 = temp.1, exp.2 = temp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', pValue = 0.01, ci = 0.90)

dysreg.res <- dysreg.out$dysreg

plotDysregExp(tf = dysreg.res$TF[1], tar = dysreg.res$Target[1],
              exp.1 = exp.1,exp.2 = exp.2, exp1.lab = 'PD',exp2.lab = 'RP',
              dysreg = dysreg.res, method ='dysreg', conf.int.level = 0.95)
              
}

}
