% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{plotDysregKM}
\alias{plotDysregKM}
\title{Kaplan-Meier plots for a individual gene dysregualtion}
\usage{
plotDysregKM(tf, tar, exp.data, surv.data, 
             divide.point = median, xlab, ylab, 
             pval = TRUE, pval.method = TRUE, 
             legend.position = c(0.8,0.9), ...)
}
\arguments{
\item{tf}{The TF of a gene dysregulation.}

\item{tar}{The target of a gene dysregulation.}

\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{surv.data}{The survival data. The surv.data  must include at least three colomns, "sample", "time", "status".}

\item{divide.point}{The point used to divide samples into subgroups.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{pval}{Logical value. If TRUE, the p-value is added on the plot.}

\item{pval.method}{Whether to add a text with the test name used for calculating the pvalue corresponding to survival curves' comparison. Used only when pval=TRUE.}

\item{legend.position}{The position for presentating legend on the plot.}

\item{...}{Other parameters passed to \link{ggsurvplot}.}
}
\description{
Divide samples into different subgroups based the expression value of TF and target of a special gene dysregulation. plot Kaplan-Meier curves between the divided subgroups.
}
\examples{
\donttest{
data(ExpData)
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]

set.seed(1234)
test.genes <- sample(1:nrow(ExpData),1000)

tmp.1 <- exp.1[test.genes,]
tmp.2 <- exp.2[test.genes,]

dysreg.out <- DysReg(exp.1 = tmp.1, exp.2 = tmp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', 
                     pValue = 0.01, ci = 0.90, verbose = T)

surv.data <- ClinData[,c("sample", "os", "censOS")]
colnames(surv.data)[2:3] <- c('time','status')

plotDysregKM(tf = dysreg.res$TF[1], tar = dysreg.res$Target[1], 
             exp.data = ExpData, surv.data = surv.data, 
             divede.point = median, 
             xlab = 'Months', ylab = 'Survival probability', 
             pval = TRUE, pval.method = TRUE,
             legend.position = c(0.8,0.9))
              
}

}
