% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{quantiReg}
\alias{quantiReg}
\title{Quantify regulatory intensities of regulations}
\usage{
quantiReg(exp, net, ci)
}
\arguments{
\item{exp}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{net}{Conditioanl GRN outpou from \code{condiGRN}.}

\item{ci}{Confident invetals of coefficients.}
}
\value{
A data frame containing the regulatory intenty and its confident invertal for each regulation.
}
\description{
Quantify regulatory intensities with de-biased LASSO.
}
\examples{
\donttest{
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

data(clin_data)
head(clin_data)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]
temp <- exp.1[sample(1:nrow(exp.1),100),]

## Build a conditional GRN based on a prior GRN
net.1 <- condiGRN(exp = temp, tf2tar = tf2tar, method = 'Boruta', pValue = 0.01)

## Quantify regulatory intensity
quanti.net.1 <- quantiReg(exp = temp, net = net.1, ci = 0.95)
}

}
