% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.2.R
\name{RankDysTF}
\alias{RankDysTF}
\title{Rank TF with dysregulatin degree}
\usage{
RankDysTF(dysreg)
}
\arguments{
\item{dysreg}{The results of \code{\link{DysReg}}.}
}
\value{
The rank of TFs.
}
\description{
Rank TF  based on dysregulation degree quantified by combining differential regulation and differential expression.
}
\examples{
\donttest{
data(ExpData)
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]


dysreg.out <- DysReg(exp.1 = exp.1, exp.2 = exp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', 
                     pValue = 0.01, ci = 0.90, verbose = T)

dysreg.res <- dysreg.out$dysreg

tf.rank <- RankDysTF(dysreg = dysreg.res)
head(tf.rank)

}

}
\seealso{
\code{\link{RankDysReg}}
}
