\name{tf2tar}
\docType{data}
\alias{tf2tar}
\title{TF-target regulations}
\description{
TF-target regulations within a reference GRN.
}
\usage{
data(tf2tar)
}
\details{
The TF-target relationships were predicted by FIMO with TF motif data, which was downloaded from HumanTF database.
}
\references{
Grant C E, Bailey T L, Noble W S. FIMO: scanning for occurrences of a given motif. Bioinformatics. 2011, 27(7): 1017-8.
Lambert S A, Jolma A, Campitelli L F, et al. The Human Transcription Factors. Cell. 2018, 172(4): 650-65.
}
\examples{
data(tf2tar)
head(tf2tar)
}
\keyword{datasets}
