% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{condiGRN}
\alias{condiGRN}
\title{Build conditional GRN}
\usage{
condiGRN(exp.data, tf2tar, 
         method = 'Boruta', pValue = 0.01, 
         threshold = NULL, verbose = TRUE, ...)
}
\arguments{
\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The prior reference GRN containing TF-target relationships.}

\item{method}{The method used to build conditional GRN, such as 'Boruta', or 'RGBM'.}

\item{pValue}{Confidence level used in Boruta. Default value should be used.}

\item{threshold}{The threshould for weight in RGBM.}

\item{verbose}{A logical value indicating whether display the computating progress.}

\item{...}{Other parameters passed to \link{Boruta} or \link{RGBM}.}
}
\value{
Conditional GRN.
}
\description{
Build conditional gene regulatory network (GRN) with expression data and prior reference network by using feature selection algorithm.
}
\details{
While using method Boruta, the predifined pValue could be used as the threshold to filter out nonsignificant regulatory relationships.  While using method RGBM, users need to explore the threshould of weight based on the output of RGBM to filter out nonsignificant regulatory relationships before following analysis.
}
\examples{
\donttest{
# Build a conditional GRN based on a reference GRN.
data(ExpData)
ExpData[1:5,1:5]

data(tf2tar)
head(tf2tar)

data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

## using method Boruta
net.1 <- condiGRN(exp.data = exp.1, tf2tar = tf2tar, method = 'Boruta', pValue = 0.01)

}

}
\references{
Kursa M B, Rudnicki W R. Feature Selection with the Boruta Package. Journal of Statistical Software. 2010, 36(11): 13.

Mall R, Cerulo L, Garofano L, et al. RGBM: regularized gradient boosting machines for identification of the transcriptional regulators of discrete glioma subtypes. Nucleic Acids Res. 2018, 46 (7), e39.
}
