% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.2.R
\name{DiffCor}
\alias{DiffCor}
\title{Differential correlation analysis}
\usage{
DiffCor(exp.1, exp.2, tf2tar, cor.method = 'pearson', p.adj, verbose = TRUE)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The  prior reference GRN containing TF-target interaction pairs.}

\item{cor.method}{The method used to calculate correlation coefficient.}

\item{p.adj}{Correction method for p value adjust.}

\item{verbose}{A logical value indicating whether display the computating progress.}
}
\value{
The identified differential regulations.
}
\description{
Differential correlation analysis with Fishers' Z test.
}
\examples{
\donttest{

data(ExpData) 
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]

## implement differential correlation analysis
diffcor.res <- DiffCor(exp.1 = exp.1, exp.2 = exp.2, tf2tar, 
                       cor.method = 'pearson', p.adj = 'BH',
                       verbose = TRUE)
                       
## set cutoff
diffcor.res <- subset(diffcor.res,p.val < 0.01)
head(diffcor.res)

}

}
