% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.2.R
\name{DiffCorPlus}
\alias{DiffCorPlus}
\title{Implement differential correlation analysis based on \code{\link{DiffCor}} by combining differential expression of target, and the consistency between correlation change and target expression change}
\usage{
DiffCorPlus(exp.1, exp.2, tf2tar, 
            de.genes = NULL, de.pval = NULL, de.qval = NULL, de.logFC = NULL, 
            cor.method = "pearson", p.adj = "BH", verbose = TRUE)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The  prior reference GRN containing TF-target interaction pairs.}

\item{de.genes}{A dataframe for differential expression genes. If no de.genes offered, DysReg uses the default method limma to implement differential expression analysis. If de.genes offered, The dataframe must include three columns, "GeneSymbol", "high.condition", "de.logFC". "high.condition" means which condition represents high expression level. "de.logFC" is the output logFC from differential expression analysis.}

\item{de.pval}{The cutoff of pval for filtering differential expression genes. If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.05.}

\item{de.qval}{The cutoff of qval used for filtering differential expression genes.  If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.05.}

\item{de.logFC}{The cutoff of absolute logFC used for filtering differential expression genes. If you don't use this parameter to filter differential expression genes, this parameter could be set as NULL. If you use this parameter to filter differential expression genes, this parameter could be set as a special number, such as 0.5. This parameter could be used by combining with de.pval or de.qval.}

\item{cor.method}{Which correlation coefficient (or covariance) is to be computed. One of "pearson" (default) or "spearman", can be abbreviated.}

\item{p.adj}{Correction method for p value adjust.}

\item{verbose}{A logical value indicating whether display the computing progress.}
}
\value{
The identified regulations.
}
\description{
Identify regulations with differential correlation, differential expression of target, and the consistency between differential correlation and differential expression.
}
\examples{
\donttest{
data(ExpData)
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

group.2 <- ClinData$sample[which(ClinData$binaryResponse == 'SD/PD')]
exp.2 <- ExpData[,colnames(ExpData) \%in\% group.2]


## implement differential correlation analysis
diffcor.p.res <- DiffCorPlus(exp.1 = exp.1,exp.2 = exp.2, tf2tar, 
                             de.genes = NULL, de.pval = 0.05, 
                             cor.method = 'pearson', p.adj = 'BH')

## set cutoff
diffcor.p.res <- subset(diffcor.p.res,p.val < 0.05)
head(diffreg.p.res)

}

}
\seealso{
\code{\link{DiffCor}}
}
