% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{DysReg}
\alias{DysReg}
\title{Identify gene dysregulations}
\usage{
DysReg(exp.1, exp.2, tf2tar, de.genes = NULL,
              de.pval = NULL, de.qval = NULL, de.logFC = NULL, 
              grn.method = 'Boruta', 
              pValue = 0.01, threshold = NULL, ci = 0.95)
}
\arguments{
\item{exp.1}{Expression matrix of a special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{exp.2}{Expression matrix of an another special condition. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{tf2tar}{The TF-target interaction pairs within a prior GRN.}

\item{de.genes}{A dataframe for differential expression genes. The dataframe must include three colomns, "GeneSymbol", "high.condition", "logFC". "high.condition" means which condition represents high expression. "logFC" is the output logFC from differential expression analysis.}

\item{de.pval}{The cutoff of pval used for filtering differential expression genes.}

\item{de.qval}{The cutoff of qval used for filtering differential expression genes.}

\item{de.logFC}{The cutoff of absolute logFC used for filtering differential expression genes.}

\item{grn.method}{The method used for constructing GRN,such as 'Boruta', 'RGBM'.}

\item{pValue}{Confidence level used in Boruta.}

\item{threshold}{The threshould for weight in RGBM.}

\item{ci}{The confident invetals of coefficients.}
}
\value{
The results of gene dysregulation analysis:
 \item{de.genes}{The identified differential expression genes between conditions.}
 \item{net.1}{The conditional GRN for exp.1.}
 \item{quanti.reg.1}{The quantification of regulatory intensity for exp.1.}
 \item{net.2}{The conditional GRN for exp.2.}
 \item{quanti.reg.2}{The quantification of regulatory intensity for exp.2.}
 \item{dysreg.res}{The identified gene dysregulations}
 
 @seealso condiGRN, quantiReg.
}
\description{
Identify gene dysregulations by integrating three properties including differential regulation, differential expression of target, and the consistency between differential regulation and differential expression.
}
\details{
DysReg first build conditional GRNs with a random forest-based feature selection algorithm, where each link??s regulatory intensity and its confidential interval is estimated with a de-biased LASSO method. Gene dysregulations were then identified by integrating three properties including differential regulation, differential expression of target, and the consistency between differential regulation and differential expression.
}
\examples{
\donttest{
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

data(clin_data)
head(clin_data)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]

group.2 <- clin_data$sample[which(clin_data$ResponseLevel \%in\% c('CR','PR','SD'))]
exp.2 <- exp[,colnames(exp) \%in\% group.2]

set.seed(1234)
test.genes <- sample(1:nrow(exp),1000)

temp.1 <- exp.1[test.genes,]
temp.2 <- exp.2[test.genes,]

dysreg.out <- DysReg(exp.1 = temp.1, exp.2 = temp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', pValue = 0.01, ci = 0.90)
}

}
