\name{ExpData}
\docType{data}
\alias{ExpData}
\title{Example of expression data}
\description{
An example of expression data.
}
\usage{
data(ExpData)
}
\details{
The example data was generated by randomly selecting 1000 genes from RNA-seq data of IMvigor210CoreBiologies.
}
\examples{
data(ExpData)
dim(ExpData)
ExpData[1:5,1:5]
}
\keyword{datasets}
