% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.3.R
\name{quantiReg}
\alias{quantiReg}
\title{Quantify regulatory intensities of regulations}
\usage{
quantiReg(exp.data, net, ci)
}
\arguments{
\item{exp.data}{Expression matrix. Columns correspond to genes, rows correspond to experiments. The matrix is expected to be already normalized.}

\item{net}{Conditioanl GRN output from \code{\link{condiGRN}}.}

\item{ci}{Confidence interval of coefficient.}
}
\value{
A data frame containing the regulatory intensity and its confidence invertal for each regulation.
}
\description{
Quantify regulatory intensities and its confidence intervals with de-biased LASSO.
}
\examples{
\donttest{
data(ExpData)
data(tf2tar)
data(ClinData)

group.1 <- ClinData$sample[which(ClinData$binaryResponse == 'CR/PR')]
exp.1 <- ExpData[,colnames(ExpData) \%in\% group.1]

## using method Boruta
net.1 <- condiGRN(exp.data = exp.1, tf2tar = tf2tar, method = 'Boruta', pValue = 0.01)

## Quantify regulatory intensity
quanti.net.1 <- quantiReg(exp.data = exp.1, net = net.1, ci = 0.90)

}

}
\references{
Javanmard A, Montanari A. Confidence intervals and hypothesis testing for high-dimensional regression. Journal of Machine Learning Research. 2014, 15(1): 2869-909.
}
