% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysReg_2.2.1.R
\name{RankDysTF}
\alias{RankDysTF}
\title{Rank TF with dysregulatin degree.}
\usage{
RankDysTF(dysreg.res)
}
\arguments{
\item{dysreg.res}{The results of \code{DysReg}.}
}
\value{
The rank of TFs.
}
\description{
Rank TF with dysregualtion degree, which is measured by differential regulation and differential expression.
}
\examples{
\donttest{
data(exp)
exp[1:5,1:5]

data(tf2tar)
head(tf2tar)

group.1 <- clin_data$sample[which(clin_data$ResponseLevel == 'PD')]
exp.1 <- exp[,colnames(exp) \%in\% group.1]

group.2 <- clin_data$sample[which(clin_data$ResponseLevel \%in\% c('CR','PR','SD'))]
exp.2 <- exp[,colnames(exp) \%in\% group.2]

set.seed(1234)
test.genes <- sample(1:nrow(exp),1000)

temp.1 <- exp.1[test.genes,]
temp.2 <- exp.2[test.genes,]

dysreg.out <- DysReg(exp.1 = temp.1, exp.2 = temp.2, tf2tar, 
                     de.genes = NULL, de.pval = 0.05, 
                     grn.method = 'Boruta', pValue = 0.01, ci = 0.90)

dysreg.res <- dysreg.out$dysreg.res

tf.rank <- RankDysTF(dysreg.res)
}

}
